Added exists service and template variable. Modified get svc arguments to be optional since they basically are anyway. Modified heredoc for get service to accurately reflect method. Increased version number from 1.0 to 1.1
Exists service added gets url for the request with the default option set
to 404. Then using guzzle we attempt to fetch the url. If a 404 status
code is returned the exception is thrown, caught and false is returend. If
no exception is thrown, then account exists and true is returned.